fix: refactor list column mixin
