+++ cosmoz-omnitable-column-list-mixin.js
@@ -1,5 +1,8 @@
 import '@webcomponents/shadycss/entrypoints/apply-shim';
 import { dedupingMixin } from '@polymer/polymer/lib/utils/mixin';
+import {
+	prop, array
+} from '@neovici/cosmoz-autocomplete/lib/utils';
 
 /**
  * @polymer
@@ -26,36 +29,79 @@ export const listColumnMixin = dedupingMixin(base =>	class extends base {
 			},
 
 			textProperty: {
-				type: String,
-				value: 'label'
+				type: String
 			},
 
 			valueProperty: {
-				type: String,
-				value: 'value'
+				type: String
+			},
+
+			_source: {
+				type: Array,
+				computed: '_computeSource(values, valueProperty, textProperty)'
 			}
 		};
 	}
 
-	getString(item, valuePath = this.valuePath) {
-		const values = this.get(valuePath, item);
-
-		if (values == null) {
-			return '';
-		}
-
-		return values
-			.map(value => this.textProperty ? this.get(this.textProperty, value) : value)
-			.filter((value, index, array) => array.indexOf(value) === index)
+	getString(item, valuePath = this.valuePath, textProperty = this.textProperty) {
+		return array(this.get(valuePath, item))
+			.map(prop(textProperty))
 			.filter(Boolean)
 			.join(', ');
+	}
+
+	getTexts(valuePath, item, textProperty = this.textProperty) {
+		return array(this.get(valuePath, item)).map(prop(textProperty));
 	}
 
 	toXlsxValue(item, valuePath = this.valuePath) {
 		return this.getString(item, valuePath);
 	}
 
+	/**
+	 * Get the comparable value of an item.
+	 *
+	 * @param {Object} item Item to be processed
+	 * @param {String} valuePath Property path
+	 * @returns {String|void} Valid value or void
+	 */
+	getComparableValue(item, valuePath) {
+		const value = super.getComparableValue(item, valuePath);
+		if (this.valueProperty == null) {
+			return value;
+		}
+		const subValues = value.reduce((acc, subItem) => {
+			acc.push(this.get(this.valueProperty, subItem));
+			return acc;
+		}, []);
+		return subValues.sort().join(' ');
+	}
+
 	_getDefaultFilter() {
 		return [];
 	}
+
+	_computeSource(values, valueProperty) {
+		return this._unique(values, valueProperty) || [];
+	}
+
+	_applyMultiFilter(filters, item) {
+		const values = array(this.get(this.valuePath, item)).map(prop(this.valueProperty));
+		return filters.some(filter => values.includes(filter));
+	}
+
+	_computeValue(
+		filter,
+		source = this._source || [],
+		valueProperty = this.valueProperty
+	) {
+		return source.filter(item =>
+			filter.includes(prop(valueProperty)(item))
+		);
+	}
+
+	_headerValueChanged({ detail: { value }}) {
+		this.filter = value.map(prop(this.valueProperty));
+	}
 });
+
