+++ cosmoz-omnitable-column-list.js
@@ -9,9 +9,6 @@ import { translatable } from '@neovici/cosmoz-i18next';
 import { columnMixin } from './cosmoz-omnitable-column-mixin';
 import { listColumnMixin } from './cosmoz-omnitable-column-list-mixin';
 import '@neovici/cosmoz-autocomplete';
-import {
-	prop, array
-} from '@neovici/cosmoz-autocomplete/lib/utils';
 
 /**
  * @polymer
@@ -66,70 +63,6 @@ class OmnitableColumnList extends listColumnMixin(
 		return 'cosmoz-omnitable-column-list';
 	}
 
-	static get properties() {
-		return {
-			_source: {
-				type: Array,
-				computed: '_computeSource(values, valueProperty, textProperty)'
-			},
-
-			textProperty: {
-				type: String,
-				value: ''
-			},
-
-			valueProperty: {
-				type: String,
-				value: ''
-			}
-		};
-	}
-
-	getTexts(valuePath, item, textProperty) {
-		return array(this.get(valuePath, item)).map(prop(textProperty));
-	}
-
-	_applyMultiFilter(filters, item) {
-		const values = array(this.get(this.valuePath, item)).map(prop(this.valueProperty));
-		return filters.some(filter => values.includes(filter));
-	}
-
-	_computeSource(values, valueProperty) {
-		return this._unique(values, valueProperty) || [];
-	}
-
-	_computeValue(
-		filter,
-		source = this._source,
-		valueProperty = this.valueProperty
-	) {
-		return source.filter(item =>
-			filter.includes(prop(valueProperty)(item))
-		);
-	}
-
-	_headerValueChanged({ detail: { value }}) {
-		this.filter = value.map(prop(this.valueProperty));
-	}
-
-	/**
-	 * Get the comparable value of an item.
-	 *
-	 * @param {Object} item Item to be processed
-	 * @param {String} valuePath Property path
-	 * @returns {String|void} Valid value or void
-	 */
-	getComparableValue(item, valuePath) {
-		const value = super.getComparableValue(item, valuePath);
-		if (this.valueProperty == null) {
-			return value;
-		}
-		const subValues = value.reduce((acc, subItem) => {
-			acc.push(this.get(this.valueProperty, subItem));
-			return acc;
-		}, []);
-		return subValues.sort().join(' ');
-	}
 }
 
 customElements.define(OmnitableColumnList.is, OmnitableColumnList);
